Move snow.response to snow.request.response
