+++ snow/request/patch.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any
 
-from snow.response import Response
-
 from .base import Request
 
 if TYPE_CHECKING:
@@ -19,7 +17,9 @@ class PatchRequest(Request):
         super(PatchRequest, self).__init__(resource)
 
     def __repr__(self) -> str:
-        return self._format_repr(f"object_id: {self.object_id}, payload: {self.payload}")
+        return self._format_repr(
+            f"object_id: {self.object_id}, payload: {self.payload}"
+        )
 
     async def send(self, *args: Any, **kwargs: Any) -> Any:
         return await self._send(data=self.payload, **kwargs)
